add: 필수값 체크 구현 [#13]
